Move inlines support to TextBlock and rename RichTextBlock to SelectableTextBlock
