+++ samples/ControlCatalog/Pages/TextBlockPage.xaml
@@ -118,7 +118,7 @@
         </StackPanel>
       </Border>
       <Border>
-        <RichTextBlock SelectionBrush="LightBlue" IsTextSelectionEnabled="True" Margin="10" TextWrapping="Wrap">
+        <SelectableTextBlock SelectionBrush="LightBlue" Margin="10" TextWrapping="Wrap">
           This <Span FontWeight="Bold">is</Span> a
           <Span Background="Silver" Foreground="Maroon">TextBlock</Span>
           with <Span TextDecorations="Underline">several</Span>
@@ -126,7 +126,7 @@
           <Span Foreground="Blue">
             using a <Bold>variety</Bold> of <Italic>styles</Italic>
           </Span>.
-        </RichTextBlock>
+        </SelectableTextBlock>
       </Border>
     </WrapPanel>
   </StackPanel>
