+++ src/Avalonia.Base/Media/TextFormatting/TextFormatterImpl.cs
@@ -502,7 +502,7 @@ namespace Avalonia.Media.TextFormatting
 
                     case { } drawableTextRun:
                         {
-                            if (currentWidth + drawableTextRun.Size.Width > paragraphWidth)
+                            if (currentWidth + drawableTextRun.Size.Width >= paragraphWidth)
                             {
                                 goto found;
                             }
@@ -665,7 +665,7 @@ namespace Avalonia.Media.TextFormatting
 
                 if (!breakFound)
                 {
-                    currentLength += currentRun.Text.Length;
+                    currentLength += currentRun.TextSourceLength;
 
                     continue;
                 }
