+++ src/Skia/Avalonia.Skia/TextShaperImpl.cs
@@ -64,7 +64,7 @@ namespace Avalonia.Skia
 
                     var glyphOffset = GetGlyphOffset(glyphPositions, i, textScale);
 
-                    if(glyphIndex == 0 && text.Buffer.Span[glyphCluster] == '\t')
+                    if(text.Buffer.Span[glyphCluster] == '\t')
                     {
                         glyphIndex = typeface.GetGlyph(' ');
